refactor: 쇼핑화면과 알림화면 리펙토링
